requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-jobspy==1.1.82
pandas==2.2.2